Separate FrameCallbackPool
